+++ Books/package.json
@@ -1,12 +1,14 @@
 {
-
   "dependencies": {
     "body-parser": "^1.20.2",
+    "chai": "^4.3.10",
+    "chai-http": "^4.4.0",
     "express": "^4.18.2"
   },
   "devDependencies": {
-    "chai": "^5.1.0",
-    "chai-http": "^4.4.0",
     "mocha": "^10.3.0"
+  },
+  "scripts": {
+    "test": "mocha"
   }
 }
